* Misc cleanup

+++ serialq.h
@@ -1,9 +1,3 @@
-// ------------------------------------------------------------------
-// --- serialq.c - basic routines for sending data over USART     ---
-// ---                                                            ---
-// ---            03.dec.2013, Matej Kogovsek (matej@hamradio.si) ---
-// ------------------------------------------------------------------
-
 #ifndef MAT_SERIALQ_H
 #define MAT_SERIALQ_H
 
@@ -18,7 +12,7 @@ uint8_t ser_getc(const uint8_t devnum, uint8_t* const d);
 
 void ser_putc(const uint8_t devnum, const char a);
 void ser_puts(const uint8_t devnum, const char* s);
-void ser_putsn(const uint8_t devnum, const char* s, uint8_t n);
+void ser_putsn(const uint8_t devnum, const char* s, uint16_t n);
 void ser_puti_lc(const uint8_t devnum, const int32_t a, const uint8_t r, uint8_t w, char c);
 void ser_putf(const uint8_t devnum, float f, uint8_t prec);
 #define ser_puti(a, b, c) ser_puti_lc(a, b, c, 0, 'x')
